Add linter to enforce that dependencies are listed in package.json
Summary:
We import several transitive dependencies at the moment.
This is extremely prone to breakage in production - if we add/remove dependencies such that the transitive dependency is now only included in dev mode, neither Flow nor lint will complain!

This adds a simple linter that checks that all `import`/`require` statements have an entry in the root `package.json`. See the previous diff for the cases that were caught - `babel-traverse` was actually throwing in production :(

I had to whitelist 2 cases that are legitimately only used in dev mode.

Reviewed By: matthewwithanm

Differential Revision: D5734614

fbshipit-source-id: b67f30aa3bd22ef4d60979f2f32a830299168073