Add support for LSP TextEdits in completions
Summary:
The LSP allows servers to use a TextEdit to describe what should happen when a CompletionItem is selected. Currently, Atom's Completion type does not support this, but we can implement it using the onDidInsertSuggestion of the AutocompleteProvider.

Forcing `autocomplete-plus` to take a suggestion as a no-op is a bit subtle, but we basically just have to empty the text, snippet, and replacementPrefix fields.

`autocomplete-plus` actually discards truly empty suggestions so we have to provide an empty snippet (`"${#:}"`) instead, where # is a number. It also de-duplicates snippets so we provide `${1:}`, `${2:}`, etc.

Reviewed By: ljw1004

Differential Revision: D5608532

fbshipit-source-id: c80e5e20a48dd80b3dd7b4f36d70bcf2fbfc0c2d