Preserve order of imports when inserting new ones
Summary:
This implements the final work necessary to make JS imports viable: preserving ordering of imports!

I decided on something sensible for FB users, at least:

1. Node modules first
2. Relative imports (../)
3. Local relative imports (./)

And within each of those we sort alphabetically, but ordering `Capitalized` strings above the rest for node modules (as is the case in our web code).

Reviewed By: wbinnssmith

Differential Revision: D5730823

fbshipit-source-id: a058d5b68b753d5a59b2a0d71d1be116dda825c6