observe .hg/progress file
Summary:
We want to utilize the progressfile hg extension to be able to show progress bars in Nuclide.

Unfortunately, setting the env option that turns off HGPLAIN for progress (and lets the extension work through Nuclide) also puts progress bars in the console.  We'll have to get the hg extension changed.

Reviewed By: mostafaeweda

Differential Revision: D5158235

fbshipit-source-id: eb1e77c0bb65c2fa84bfb13d3a87a0f28a8d5d1a