Fix PHP Debugger giving EISDIR red box during debugging
Summary:
The script ID being sent by HHVM for the top frame in this case is actually a directory, not a file. I am not sure why it is doing that, but that is external to Nuclide and we should handle it more gracefully.

In this case, the file cache is throwing trying to read the file. I'm fairly certain this didn't work before either, but the change to VS Code adapter lit up this path.

Reviewed By: mostafaeweda

Differential Revision: D5747462

fbshipit-source-id: 0d5c45b5e75fac7cc54720dcbc5a86dcb8c02c83