Watch directories with node
Summary: `fs.watch` supports watching both files and directories. The event callback returns a relative path to the file changed - unfortunately Nuclide's RPC marshaling converts the change's relative path to a remote URL, dropping the portion of the original watched path. For files we can keep doing what we were doing and assume the watched path has changed on any change. For directories we need to join the watched path with the new relative one. Asking the client whether they want to watch a directory or not isn't ideal, but we don't have too many use cases for this yet anyway.

Reviewed By: hansonw

Differential Revision: D5741811

fbshipit-source-id: e7fffad66ffdf342b201fe08f4fcb2553c1ccb80