De-clown java debugger port selection
Summary:
Currently Java debugger port is hard-coded to 12345. Aside from being a silly number, this is causing debugging to fail if another process is using that port. Changing to dynamically select an available port on the remote host.

Some process on my dev server is hung and using this port (not even sure which one) and short of rebooting I can't Java debug now. Fixing this for all cases.

Reviewed By: mostafaeweda

Differential Revision: D5758739

fbshipit-source-id: 7a7c08494f0eb5c5eef0159fac203c1641dbefdd