Hyperclick: avoid duplicate fetching with multiple keydown events
Summary:
If the current position is equal to the last position and we haven't gotten a result back, that's perfectly fine - we should wait for the previous promise to finish.

This is meant to address a particularly egregious case on Windows where modifier keys also repeat! So if you hold "Ctrl" it seems that multiple keydown events get emitted.. which causes a lot of extra fetching for Hyperclick.

I'm not sure if 'keyup' events are also fired, but it seems like that's OK.

https://github.com/facebooknuclide/hyperclick/issues/30

Reviewed By: matthewwithanm

Differential Revision: D5777577

fbshipit-source-id: 05991de4d7efc9d464271a8a30d39b6de483142f