Move debugger:show earlier in start debugging
Summary:
Currently the last thing we are doing when starting debugging is showing the debugger pane if it's not already visible. Sometimes it takes a while for the debugger backend to start up and attach to the target, which is not noticable if the debugger pane happens to already be open, but if it's not it looks like nothing is happening.

This change opens the pane before starting the backend so that the user can see the "Starting debugger..." message in the debugger pane and know we haven't failed.

Reviewed By: mostafaeweda

Differential Revision: D5781248

fbshipit-source-id: 7c3f4870b6b8b13f7dc935328250eb29a9f2ddee