Fix sdb configuration for device listing
Summary:
Refactored code to utilize the fact that the sdb framework doesn't need to explicitly use a port number in its commands like the sdb. Also, the port number isn't necessary for the device lisitng in either case. So I took the liberty of cleaning up some code.

This is the first part of the task.

Reviewed By: a20012251

Differential Revision: D5777567

fbshipit-source-id: 5db2d472ae45bcb4eddce0f523729d446f9963d4