"Toggle Blame" command works on active text editor even if not focused
Summary:
You can now Toggle Blame for the active text editor (both from the Command Palette and the context menu) even if you aren't focused on the text editor. If there is no focused text editor, we instead look for the workspace's center's active item and if that is an editor we use that instead.

Also added 'Nuclide > Blame > Toggle' and 'Nuclide > Blame > Hide' to the Atom menus:
https://pxl.cl/8P59

Reviewed By: matthewwithanm

Differential Revision: D5758867

fbshipit-source-id: f56d316990c0f2e7cb41036af94bef07cc1a7e9a