Cancel textDocument/definition requests
Summary:
Cancel old definition requests when the new one is sent.

Note that a definition request is sent whenever you press command on a buffer. I do this all the time, because I command-tab all the time. This makes OCaml unusable, because every definition request takes 5+ seconds. This is like a breath of fresh air in the Seattle smog. Why didn't I do this before?

There is a question of what this means for languages which are not OCaml. We issue `textDocument/definition` to resolve the definition location of a symbol, which happens in two situations:

* When we right-click and select "go to declaration".
* When we command-click or otherwise trigger Hyperclick™ or whatever it's called. When we hover over a symbol with command pressed, we issue a definition request. When the result is received, we underline the symbol at the current position to indicate that we can jump to it.

Conceivably there could be other cases of requesting the definition of a symbol in the editor in the future, for which it doesn't make sense to cancel this request. But I don't think that's likely.

Right now, I think it makes sense for all languages to implement this cancellation behavior. If I decide I want to go to another symbol, I want my most recent request to be honored, and not go through a bunch of intermediate files as the language server resolves them.

Reviewed By: ljw1004

Differential Revision: D5779956

fbshipit-source-id: 3721b4d057cabb7b11c038dd1919150c41ff5d09