Add did save support to LSP
Summary:
Added did_save support to the LSP language service. I wasn't sure what the best
way to include the buffer text was, so it's null for now? Should we grab the contents directly from
the buffer, or rely on the FileCache for the correct contents? Also, I am not very familiar with the
codebase so most changes were done via imitating existing structures, and would appreciate comments on style/better
ways to accomplish things in JS.

Reviewed By: hansonw

Differential Revision: D5769894

fbshipit-source-id: 7d02e64164afdf5752ca00f2e707eb9b5a5d3de3