refactor SshHandshake::connect
Summary:
Made `SshClient::connect` handle 'read' and 'error' events and return a `Promise`, and then replaced the event-listening logic in `SshHandshake` with `Promise`-based logic. The goal was to simplify the logic (make it local) and to resolve a bug where connections do not properly fail if an error occurs.

Depends on D5729387

Reviewed By: bolinfest

Differential Revision: D5734328

fbshipit-source-id: 001c0e450665c7f7f20e0e0980bdf6b3a6abe341