Support displayNames for features in Settings
Summary:
This standardizes on a `displayName` field in a feature's
`package.json` (happy to change the field) as a way of displaying a
human-friendly title of a feature.

These friendly names are used when generating each feature's section in
settings, so it doesn't end up as a capitalized hyphenated identifier.

This also uses these names in the "Enabled Packages" ('use') section.

Reviewed By: matthewwithanm

Differential Revision: D5832001

fbshipit-source-id: 37ec8acf0349f81bf82ba3313fad0226d805006a