add ZIP/JAR support (4 of 4): integrate to FileTree/FileSystemService
Summary:
This integrates the FileSystem abstraction into the rest of Nuclide.  This includes FileTree, FileSystemService, and a couple changes in the remote buffer management.

I ended up adding support for archive paths directly in the remote file handler because that includes a bunch of code to coordinate with Atom at various versions.  Once we remove support for old versions of Atom, we might consider factoring out the coordinating code and using it from a separate addOpener that is specific to local archive paths.

Reviewed By: shushz

Differential Revision: D5795380

fbshipit-source-id: c01cca3818ab9d5a2b0bb95e9fc2e6879ed33bce