Enable flow typechecks on buffer code
Summary: Nuclide had to have separate code paths before and after Atom v1.19, which made things hard to typecheck.  Now that Atom v1.19 has had a chance to roll out, this removes the old codepath and adds the typechecks.  This came up during a review of archive file support, and fixes an issue there with createReadStream to fit the typecheck.

Reviewed By: hansonw

Differential Revision: D5832200

fbshipit-source-id: de4d22c6652c662927fdf30f6dc2651825e94827