Add config option to show debugger if hidden when setting breakpoint
Summary:
This change pops open the debugger when setting a breakpoint in the gutter. Added a config option to disable this, since some users might like it the other way. T11151919 suggested showing a notification that you need to start debugging, but I think now that the debugger pane indicates the debugger isn't attached, and offers you launch/attach buttons, it's better to open it up this way.

Hopefully this will increase debugger visibility and reduce confusion around setting unresolved breakpoints when no target is being debugged.

Reviewed By: mostafaeweda

Differential Revision: D5829911

fbshipit-source-id: 2e52e6f22eef76b9b476b4211572ebf8c3637621