File Tree: Switch from Polling to Resize Observer
Summary:
The file tree uses windowing to minimize the amount of DOM nodes it needs to create. Awesome! But in order to do that, it needs to know the size of the area it needs to fill, and that means knowing when the size changes. Historically, we had no way of knowing that for sure so we resorted to polling.

This diff replaces both 'resize' listener and polling for the new ResizeObservable.

Reviewed By: matthewwithanm

Differential Revision: D5853805

fbshipit-source-id: 13febd2cf21d770b7e4e51761a6bc789b33fba4c