Hack to fix remote buffer serialization on Windows
Summary:
On Windows, it appears that Atom no longer swallows EINVAL errors from attempting to open `nuclide:\a\b\c` on startup (this is an invalid drive identifier). Unfortunately this has the side effect of torpedoing the entire deserialization phase! :(

To work around this for now, we're going to have to monkey-patch the TextBuffer serialization function to sanitize these remote URIs. For now I've decided to replace `nuclide:/` with `nuclide_\` on Windows-only to make sure that the deserialized path is at least valid.

I'm working to try and find the regression within `superstring` that caused this so we can avoid this. Alternatively there's https://github.com/atom/atom/pull/15681 which may make our deserialization experience better as well.

Reviewed By: ebluestein

Differential Revision: D5868633

fbshipit-source-id: 0a09ed468380edb9d17228420fe413dbf200424d