Fixed Nuclide console autoscrolling and new messages button to scroll and stick to the bottom
Summary: Added a boolean _isScrollingToBottom to Console.js, which is used when _scrollToBottom() is called. Essentially, the log output optimizes message loading, so when you scroll to the bottom, more messages could be loaded underneath it, causing the issue where scroll does not stick to the bottom. With this boolean, it will keep calling scrollToBottom until the scrollbar is actually at the bottom.

Reviewed By: matthewwithanm

Differential Revision: D5863096

fbshipit-source-id: 6a67af061d39fc0a210dab5ce8d5742f17434622