Build using buck @mode/dbg by default when debugging
Summary: Users keep building optimized code with no symbols and then complaining that the native LLDB debugger doesn't work for their binary. This change builds mode/dbg by default when the actions is Debug for a native target - unless the user explicitly includes an mode in the build arguments from the buck toolbar.

Reviewed By: a20012251

Differential Revision: D5883075

fbshipit-source-id: ab53d9fb5339f778b6583fdf5c4b0b693ac4aacd