handle `continue` in SshClient and SftpClient.
Summary: The semantics of the `continue` event and the boolean returned from many of the ssh2 functions corresponds to `readable.push()`. When `false` is returned, it means the outgoing/readable buffer is full and that we should wait for the `continue` event (indicating that it has been emptied somewhat) before adding more data by issuing requests. Up til now, this had been ignored. And I noticed that the connection would sometimes break when transferring many files in parallel. To address this, I have added logic to `SshClient` and `SftpClient` to detect when the buffer is full, and if so, to `await` the `continue` event before issuing a new request.

Reviewed By: hansonw

Differential Revision: D5874191

fbshipit-source-id: 0d899b81e883abacfaaca5c7b53efb5a80867475