Add ability to launch a PHP debugger script in the terminal - Server side - 4/N
Summary: Implements the server side of letting a PHP script launch under the terminal. Instead of launching the script immediately from the connection multiplexer, if the deferLaunch config flag is set, the debugger just sets itself up and launches the dummy. The Nuclide front-end will be responsible for spawning the child process script.

Reviewed By: rightparen

Differential Revision: D5869319

fbshipit-source-id: 5408a41ab027d34162177445a8bb0ea632889013