Add ability to launch a PHP debugger script in the terminal - Client Side - 3/N
Summary: The client side of this change adds a flag that will set up the PHP debugger but not actually launch the target script. Instead, it calls into the newly exposed nuclide-debugger.remote service routine after the debugger is ready to launch the PHP target and have it connect to the debugger.

Reviewed By: mostafaeweda

Differential Revision: D5869297

fbshipit-source-id: 82b0e8e684df83d3ae919101d8bae0b4a9f97aca