PHP debugger: Add the ability to pass arguments to Scripts from the HHVM toolbar
Summary: This change adds a gear button that looks and behaves similar to the one we use in the Buck task runner toolbar when a PHP script is selected as the current file. Clicking the gear pops open a modal dialog that lets the user specify arguments to be passed to the script when launching it for debugging.

Reviewed By: shushz

Differential Revision: D5873861

fbshipit-source-id: 72a32d0ccff62fabb4549bd26e19bcefd5d1b4b0