Diagnostics: Don't focus the editor on single click
Summary:
We're going to try to be more consistent about this, with the file tree, outline view, and diagnostics all navigating the editor on both single and double-click, but only double-click changing focus. This will play better with keyboard navigation. Although the diagnostics table isn't very keyboard navigable, I'm working on a new UI that is.

The file tree part has already been implemented in D5798036.

Reviewed By: hansonw

Differential Revision: D5887669

fbshipit-source-id: 40d37da75675cee07fe3f9d80b98b3518a3c3af7