Only focus editor on double click
Summary:
This stops the editor from being focused on a single click of
an item, allowing easier keyboard navigation in the future and aligning
the behavior with the rest of Nuclide.

Instead, focus on double click. **This removes the old behavior of
selecting the symbol region on double click**.

If there's feedback, maybe this can return as a triple click? There's
actually a good analog in macOS/Windows text selection of a whole line
on triple click.

Reviewed By: matthewwithanm

Differential Revision: D5891757

fbshipit-source-id: 2731563f78b74a254e3a66a0459c8d3335711549