give more detailed exception callstacks
Summary:
Consider this code
   try {await p;} catch (e) {console.log(e.stack)}

If someone rejected the promise by calling `reject(new Error())`, then `e.stack` will show the callstack of that *rejector*, and you won't get any information about the *awaiter*.

This diff changes _logLspException to add in the callstack of the awaiter.

It's motivated by a frustrating set of logs where I can see that vscode-jsonrpc has encountered an EOF when writing a message to the LSP server, and then everyone who was awaiting upon jsonrpc gets an exception 'Connection got disposed', but the logs only show the callstack associated with writing to the LSP server -- they don't show the callstack where we were awaiting upon jsonrpc.

Reviewed By: hansonw

Differential Revision: D5895379

fbshipit-source-id: ea7607a6c4d66585782eba275fa59d53af7c2da7