Parse all Haste settings from .flowconfig files (1/n)
Summary:
So it turns out Haste resolution is surprisingly tricky, but we need to get it right to provide accurate suggestions.

`.flowconfig` contain several Haste settings, in fact:

1) `module.system=haste` indicates that Haste modules are enabled.
2) `module.system.haste.use_name_reducers=true` indicates that name reducers are enabled.
3) `module.system.haste.name_reducers=..` define `regex -> replacement` reduce patterns. (if not provided, the default reducer is just the basename minus extension).
4) `module.system.haste.paths.blacklist` define files blacklisted for name reducing.
5) `module.system.haste.paths.whitelist` define files whitelisted for name reducing (if not provided, this defaults to all files)

This implements the parsing in the JS imports code.

Our previous interpretation of the blacklist values was incorrect, so I just removed the handlers for now.

Reviewed By: wbinnssmith

Differential Revision: D5886126

fbshipit-source-id: 23cc279c15db86a09272cb212056ee9dfaa2bb22