Use real Haste module names for import suggestions [3/n]
Summary:
Previously, we would just take the basename, strip the extension, and assume that was the Haste module name.
Instead, for every export we'll store the actual resolved Haste name (using the name reducers or providesModule).

This took a bit of finagling to thread the `HasteSettings` object through to all the places that needed it.
I'll try and refactor this code at some point so the state can be an instance property or something.

Reviewed By: wbinnssmith

Differential Revision: D5886127

fbshipit-source-id: a81181c120f66296b71470723f8e57605111a2e5