Partially support the process table for the pixel phone
Summary:
I'm adding timeouts to the inner observables that are used to create the process table so that even with partial information the table can be displayed.
The problem is that, on the pixel phone, it's take tooo long to read the cpu and mem usage from the /proc files. It takes more than 20 seconds. I've noticed that `top` only fetches that information from the processes that are displayed on the shell at that time. I'll see how to improve that later.

Reviewed By: ebluestein

Differential Revision: D5888835

fbshipit-source-id: 5ccb0ce1ab676560a6b724e2fb65ffb9fe43595a