Rxify trackStalls with PerformanceObservable
Summary:
This exercises `PerformanceObservable` from `observable-dom`
and cleans up the stall tracking code a bit by removing all the state we
tracked manually and using a series of maps and filters on an rx
observable.

This could have gone further, but I tried to balance this with
readability. For example, `intentionalBlock` could be a stream of block
times `combineLatest`ed with the performance entries to filter there as
well. That `intentionalBlock` would itself be a merge of a stream adapted from
the atom `onWillDestroyPaneItem`, but that feels unnecessarily complex
and a lot to understand.

Reviewed By: hansonw

Differential Revision: D5906920

fbshipit-source-id: c23671f78c1f29222a7b9d44c3e4f131517954cd