Make file-tree selection management faster
Summary:
Moves the properties "isSelected", "isFocused" and "isTracked" out of the immutable part of the FileTreeNode
These properties are updated very often in performance-critical flows and having them in the immutable state causes a lot of thrashing.

With this change in place the file tree does not trigger its state recalculation on the most common of updates and common file operations. I.e. scrolling with keyboard in the file-tree up and down does not invalidate the immutable state tree. Switching tabs with "Reveal on Switch" does not either if the files that the file tree is switching to are expanded already. If not, only the first expand costs something (still pretty low).

Reviewed By: mostafaeweda

Differential Revision: D5889859

fbshipit-source-id: 5ec9e378e8599c1b4b632886ddacc26f1cb942b5