Allow files with '!' that are not archive separators
Summary: We use '!' as a separator in nuclideUri to indicate when a path refers to a file inside an archive.  We were over-aggressively verifying that a path did not end with '!' to ensure the rest of Nuclide is handling paths correctly.  We already normally ignore '!' characters that do not appear after a known archive extension.  This fixes the verification to use the same check, and adds unit tests involving a path that ends in '!'.

Reviewed By: hansonw

Differential Revision: D5921975

fbshipit-source-id: afc47beb3a601b8521555a76daaeb2fad5e6cac0