Use dropdown to select debugger connection
Summary:
When launching or attaching a debugger, use a dropdown instead of a hyperlink to select which connection (local or remote) to use.

Known issue: by default, no option is selected and it visually says "localhost" although it is actually defaulting to the remote connection. Once either option is selected it displays correctly. This was an existing issue, will fix in separate diff.

Reviewed By: ebluestein

Differential Revision: D5901884

fbshipit-source-id: ab5a3fc4447fa043b927297cdd090411ef31d88f