PHP Debugger: fix scopes not updating after an evaluation with side effects
Summary: When issuing a PHP evaluation command in the console, if it has side effects (say `$x = 5;`), the scopes pane should update to reflect the new variable value. Currently it is not updating until the frame is refreshed, which happens on the next resume+break, or if you click on the frame in the callstack pane again.

Reviewed By: Goom11

Differential Revision: D5923294

fbshipit-source-id: 63e9f69ab65c454acb7876c816976b9cc90c717f