Fix single-instancing and focus when using Terminal for debugger
Summary:
When using the terminal for the debugger, I want a single terminal pane to be reused each time you launch the debugger for the same target (otherwise they pile up and it's super annoying to the user). The URI of the terminal can change between runs of the same target if the arguments change, but we still want to use the same pane. This updates the single instancing logic in the RemoteControlService to key only off of the terminal's title, and the target executable. It will destroy existing terminals that match before creating a new one.

Additionally, I'm adding the ability to blur the terminal's focus after spawning. When you create this via the debugger, you want to be able to use debugger hotkeys like F8 to resume immediately, and if the terminal grabs focus it ruins that.

Reviewed By: rightparen

Differential Revision: D5897563

fbshipit-source-id: eb5c6640fcf3bd0b466c7783bf4a3d3a9022b79a