Support launching native debug targets in the Nuclide terminal
Summary:
This change will allow launching of a native binary under the debugger using the Terminal where available. The user can opt out of this via a new config setting.

When launching in the terminal, STDIN and STDOUT will be redirected to the terminal, so the user can interact with the target app in addition to interacting with the debugger engine (which is still done via the Console pane).

For Native targets, we support configuring the CWD as well as passing environment variables for the terminal to use when launching.

Since Linux doesn't have a mechanism to create process suspended, if we just launch the target directly and then try to attach LLDB, there is a race where we risk missing initial breakpoints set early in the program's startup, this would not be acceptable. To make this work properly, a wrapper script and some synchronization is needed to ensure LLDB is attached to the process before it gets to start executing.

The synchronization flow in this diff works like this:

1. Nuclide front-end (from LaunchProcessInfo) calls prepareForTerminalLaunch on the Nuclide native debugger RPC service
2. The backend determines the proper paths for python, the wrapper (launch.py in this diff) and the executable.
3. The backend spawns a TCP socket server that listens on a new port and waits.
4. The backend returns the values from step #2, and the port from step #3 to the front-end as launch instructions.
5. The front-end launches the wrapper script *in a new terminal pane* using the path and args the back end gave it.
6. The wrapper script connects to the listening TCP socket to talk to the debugger back end, and writes its own PID into the socke and then issues a read on the TCP socket - *this causes it to block*
7. The debugger backend receives the PID, attaches LLDB to the specified PID and then writes to the TCP socket to unblock the wrapper
8. The wrapper is unblocked from TCP read and issues an execv() syscall to replace itself with the target image, which is now free to begin execution with LLDB attached.
9. The backend closes the TCP sockets and cleans up the server.
... debugging continues as normal from here on.

Reviewed By: rightparen

Differential Revision: D5898651

fbshipit-source-id: 1c48b7eff53e0c9a2e665a42fe0de2f884c88b5b