Add ability for debugger front ends to filter breaks
Summary: When launching the native debugger in the terminal, we encounter two additional startup breaks after the expected loader breakpoint. One is a signal from the execv syscall, and the other is a SIGWINCH. This adds the ability for the individual debugger front end to be notified of a break, and given the opportunity to filter it, which would cause an auto-resume of the target.

Reviewed By: mostafaeweda

Differential Revision: D5899736

fbshipit-source-id: 79bdcf7bd03269020eae889a0357490b783cdd75