Fixed expanding properties in debugger always scrolling to the bottom of the object.
Summary: When expanding and collapsing properties of an object in console, it will always scroll to the bottom of the object. This behaviour is bad when you have a massive object and expanded a property near the top of the object. You would have to scroll up to see the expanded contents again since the scroll would have scrolled all the way to the bottom of the object.

Reviewed By: wbinnssmith

Differential Revision: D5917142

fbshipit-source-id: 6a77058f7110c2d10197b1db76f9ee361d99fcec