Datatips: Fix improperly sized datatips
Summary:
Datatips frequently display a 12px wide editor. I think I've narrowed this down to a race condition with Atom's text editors rendering things asynchronously - in particular, marker rendering is now asynchronous as well.

In the context of datatips, the React-mounted `AtomTextEditor` needs to be attached to the DOM to size themselves. Since markers are now asynchronous, the marker (and thus datatip) may not be attached by the time the React rendering occurs.

The fix seems to be waiting for the next editor update to render the datatip. At this point, the marker should be attached. It feels a bit sketchy to make some of these areas asynchronous so I added some conservative destroyed checks.

User report: https://github.com/facebook-atom/atom-ide-ui/issues/22

Reviewed By: matthewwithanm

Differential Revision: D5929694

fbshipit-source-id: a2724c85c7c0ae8ffd9d4d5d79a50c6d1ad45e38