Don't let auto-fold imports to mess on-screen position
Summary:
The auto-fold of imports comes after a slight delay when opening a file or cmd-clicking to a symbol in a new file. Unfortunately this auto-fold used to shift the visual position.
Not any more!

For some reason, the problem was especially obvious with third party libs and attached sources.

Differential Revision: D5939261

fbshipit-source-id: 99fb20095a999023bb2fe54af199fe7602b1d33d