on-demand logs 1 - rip out old additionalFiles
Summary:
This gets rid of the additionalFiles mechanism. Higher up in the stack I'm going to replace it with something more flexible.

Incidentally, observe that the only place to use additionalFiles is the Java Debugger. And it only ever uses them from the local machine - it never fetches them from a remote server.

Reviewed By: hansonw

Differential Revision: D5937645

fbshipit-source-id: 0f1969cf315f698a9a1385b66dcd9e507e71849c