on-demand logs 2 - providers of additionalLogFiles
Summary:
This is the new additionalLogFiles mechanism.

(1) It switches to "on-demand" logs that are generated at the time the user hits the bug button. This way we don't need to worry if it's expensive to gather up all the logs we want.

(2) It uses a standard Atom consumer/producer, so any package is able to offer up on-demand logs.

Reviewed By: hansonw

Differential Revision: D5937703

fbshipit-source-id: 0d77706464f06d83201d3c8e18dbf9cf740f30b4