on-demand logs 4 - HgService provides 'diff' in logs
Summary:
Something that's been really hard, when folks report bugs like "hyperclick isn't taking me to the right place" or "this error doesn't show up", is how to reproduce the error.

This takes us a small step closer to reproducibility. It offers up additional logs that let someone get back to the same mercurial "working directory" state as when the bug was filed.

Reviewed By: mostafaeweda

Differential Revision: D5937800

fbshipit-source-id: b64846b927424f4d6e651a7f27bf4d57e5c3b860