on-demand logs 6 - LSP traces
Summary:
I've been struggling how to get good logs. Previously when folks reported bugs, we had to go back to them and ask "could you please set logLevel to TRACE and then try to repro it?". But it was hard.

With this diff, LSP now keeps a transient in-memory verbose log over the past five minutes. The log is only retained if some package asks for on-demand logs.

Another trouble I had was not knowing what the contents of files were. It was too hard to reconstruct them from didOpen/didChange notifications.

With this diff, LSP now keeps transient in-memory snapshots of files over the past five minutes. It keeps snapshots no more frequently than once every 30 seconds, and only of files that the user has changed. I hope that, because of this, the overhead of such logs won't be too high.

Reviewed By: hansonw

Differential Revision: D5937934

fbshipit-source-id: 45eac6fc0f0a55dfeaaed94f225540541d43fc30