on-demand logs 8 - LSP server can rage too
Summary:
This diff proposes a new request for LSP: 'telemetry/rage'. The server indicates it can fulfill this request with a server-capabability `rageProvider?: boolean`. If the client makes this rage request, the server will respond with an array of `{title:string, data:string}` items.

It is designed like all 'rage' commands for general purpose of gathering logs and reporting bugs.

Reviewed By: hansonw

Differential Revision: D5945315

fbshipit-source-id: 7a05b911ecd0fd8545678e9318e96a0588a28fef