Disable context menu in the open files section
Summary:
The target nodes of the context menu in the open files section are determined by looking at the file tee's store, but this store serves most and foremost the file tree itself. Thus it is not guaranteed to be in sync or even contain the nodes representing the open files. This is especially visible immediately after a reload. So, basing on this is inherently wrong and error prone.

Will need to see how to address that.

Reviewed By: aadisriram

Differential Revision: D5955418

fbshipit-source-id: 946cf052892628e6c287594a02350c4672f8bc6f