Add debugger disassembly view for Native debugger
Summary:
This change adds a new disassembly view for the debugger. Individual debuggers can opt in, and currently only Native is being opted in in this diff.

This also modifies the debugger layout manager to allow a new pane to have a default location other than the right dock, so that the disassembly view can be at the bottom by default.

Reviewed By: shushz

Differential Revision: D5915585

fbshipit-source-id: 6003b597180107a4b05b772447cf0dc95014c42e