Measure the scroller height in the file tree, not just the parent component
Summary:
We should monitor resize events not just those that are coming from the parent element, but also the element that is actually containing the file tree (as it is the one being measured).
This scroller element is not necessarily in DOM at all times. For example, when the Folders section is collapsed it will not be there.

Reviewed By: matthewwithanm

Differential Revision: D5959972

fbshipit-source-id: a0d547f7812e0e6d1daeaddbd3103c9a53cb5687